+++ src/Hacks/Universal/ShowTouches/TouchTrail.cpp
@@ -2,21 +2,21 @@
 #include "Colours.hpp"
 #include <Speedhack.hpp>
 
-bool CCTouchTrail::init(CCTouch* touch)
+using namespace geode::prelude;
+using namespace qolmod;
+
+bool TouchTrail::init()
 {
     if (!CCNode::init())
         return false;
 
-    touches.emplace(touch, this);
-    
-    this->touch = touch;
     this->scheduleUpdate();
 
     streak = CCMotionStreak::create(0.3f, 2, 10, ccc3(255, 255, 255), "streak_05_001.png");
     streak->setUserObject("ignore-longer-trail"_spr, CCBool::create(true));
     streak->unscheduleUpdate();
 
-    circle = CCSprite::createWithSpriteFrameName("d_circle_02_001.png");    
+    circle = CCSprite::createWithSpriteFrameName("d_circle_02_001.png");
     circle->setOpacity(150);
 
     auto clip = CCClippingNode::create(circle);
@@ -27,24 +27,20 @@ bool CCTouchTrail::init(CCTouch* touch)
 
     durationLabel = CCLabelBMFont::create("", "bigFont.fnt");
 
-    update(0);
-
     this->addChild(clip);
     this->addChild(circle);
     this->addChild(durationLabel);
     return true;
 }
 
-void CCTouchTrail::update(float dt)
+void TouchTrail::update(float dt)
 {
     dt = Speedhack::get()->getRealDeltaTime();
     duration += dt;
 
-    auto pos = touch->getLocation();
-
-    streak->setPosition(pos);
-    circle->setPosition(pos);
-    durationLabel->setPosition(pos);
+    streak->setPosition(position);
+    circle->setPosition(position);
+    durationLabel->setPosition(position);
 
     circle->setColor(CircleColour::get()->getColour());
     streak->setColor(TrailColour::get()->getColour());
@@ -60,24 +56,7 @@ void CCTouchTrail::update(float dt)
     durationLabel->setString(fmt::format("{:.2f}", duration).c_str());
 }
 
-CCTouchTrail* CCTouchTrail::create(CCTouch* touch)
+void TouchTrail::setPosition(const CCPoint &position)
 {
-    auto pRet = new CCTouchTrail();
-
-    pRet->init(touch);
-
-    return pRet;
-}
-
-void CCTouchTrail::remove(CCTouch* touch)
-{
-    if (!touches.contains(touch))
-        return;
-
-    auto t = touches[touch];
-    t->removeFromParent();
-
-    touches.erase(touch);
-
-    delete t;
+    this->position = position;
 }